creation of pyQtWidgets.py
moved compass class into pyQtWidgets.py and preparation to merge with pyQt branch
